requirements.txt
orcid
python-dotenv
pybtex
requests
pypandoc
pandoc
toolz